move Curation to CurationPage
