+++ website/src/components/CurationPage.vue
